Add files via upload
I worked on adding a second layer of text behind the headline to have the number that it is shadowed behind it similar to the "The Information" newsletter. I am having trouble with spacing because different headline lengths alter the spacing of the number. Making it independent of the text/headline would be a good next step to work on.